headless/backend/requirements.txt
asgiref==3.8.1
build==1.2.1
click==8.1.7
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
importlib_metadata==7.1.0
packaging==24.0
psycopg2==2.9.9
pyproject_hooks==1.0.0
python-dotenv==1.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
zipp==3.18.1
